+++ packages/vui/src/components/VTextField/VTextField.tsx
@@ -1,11 +1,11 @@
 import './VTextField.scss';
 import { defineComponent, computed } from 'vue';
 import {
-  createTextInputSettings,
-  useTextInputControl,
+  createTextInputNodeSettings,
+  useTextInputNodeControl,
   createFormControlProps,
   FormControlSlots,
-  TextInputEmits,
+  TextInputNodeEmits,
 } from '@fastkit/vue-form-control';
 import { defineSlots, ExtractPropInput } from '@fastkit/vue-utils';
 import { VFormControl } from '../VFormControl';
@@ -23,7 +23,7 @@ import {
 import { VTextCounter } from '../VTextCounter';
 import { VUI_TEXT_FIELD_SYMBOL } from '../../injections';
 
-const { props, emits } = createTextInputSettings();
+const { props, emits } = createTextInputNodeSettings();
 
 const slots = defineSlots<FormControlSlots & InputBoxSlots>();
 
@@ -40,7 +40,7 @@ function createTextFieldProps() {
 
 export type TextFieldPropsOptions = ReturnType<typeof createTextFieldProps>;
 
-export type TextFieldEmits = TextInputEmits;
+export type TextFieldEmits = TextInputNodeEmits;
 
 export type TextFieldInput = ExtractPropInput<
   ReturnType<typeof createTextFieldProps>
@@ -52,7 +52,7 @@ export const VTextField = defineComponent({
   emits,
   slots,
   setup(props, ctx) {
-    const inputControl = useTextInputControl(props, ctx, {
+    const inputControl = useTextInputNodeControl(props, ctx, {
       nodeType: VUI_TEXT_FIELD_SYMBOL,
     });
     const control = useControl(props);