+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 1.1.90
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.17.0
+
 ## 1.1.89
 
 ### Patch Changes